[BABY-6] Graceful shutdown on SIGINT
